fix(turbo): bf16 output for TurboFlash pass2 kernels
Switches turbo_flash_pass2 and turbo_flash_pass2_fused_rot to write
`device bfloat* output` instead of `device float*`. Internal accumulators
(m, l, o[]) stay fp32 — only the final write narrows. Same pattern MLX
SDPA uses.

This eliminates fusion bug #87/#92 (`!!!!!` decoding on Qwen3.5-9B
nKVH=4 and similar shapes) at the source. PR #104's
`output.asType(queries.dtype)` workaround in the Swift caller is no
longer needed and is removed in the companion mlx-swift-lm commit.

Validated coherent output on Qwen 0.8B (116.5 tok/s), Qwen 9B (42.0
tok/s), Nemotron 30B (61.1 tok/s) at 4k turbo4v2 with
useCompressedAttention=true and no asType cast.

Also adds env-gated MLX_HAZARD_TRACE=1 diagnostic logging in
CommandEncoder (zero cost when disabled). Used to investigate this
bug; kept as durable infrastructure for future hazard-tracking
debugging.

+++ mlx/fast.cpp
@@ -1637,6 +1637,12 @@ std::vector<array> turbo_flash_pass1_nr0_causal(
        array(q_offset, int32)});
 }
 
+// Output is bfloat16 — matches model dtype for all current MLX LMs and avoids
+// an fp32→bf16 conversion in the consumer's graph. Kernel accumulators stay
+// fp32; only the final write narrows. This dtype change is also what fixes
+// upstream MLX graph-compile fusion bug (#87/#92) that produced `!!!!!`
+// decoding on certain shapes — see kernel comment in
+// backend/metal/kernels/turbo_quant.metal.
 array turbo_flash_pass2(
     const array& o_partials,
     const array& m_partials,
@@ -1654,7 +1660,7 @@ array turbo_flash_pass2(
 
   return array(
       {total_q, dim},
-      float32,
+      bfloat16,
       std::make_shared<TurboFlashPass2>(
           s, fallback, dim, /*fused_rotation=*/false),
       {astype(o_partials, float32, s),
@@ -1680,7 +1686,7 @@ array turbo_flash_pass2_fused(
 
   return array(
       {total_q, dim},
-      float32,
+      bfloat16,
       std::make_shared<TurboFlashPass2>(
           s, fallback, dim, /*fused_rotation=*/true),
       {astype(o_partials, float32, s),
